+++ projects/ngx-forms/src/directives/ngx-input.directive.ts
@@ -1,12 +1,10 @@
-import {
-  booleanAttribute, computed, Directive, effect, ElementRef, forwardRef, inject, input, InputSignal,
-  InputSignalWithTransform, model, ModelSignal, NgZone, OnDestroy, output, signal
-} from '@angular/core';
-import {FormNode} from "@juulsgaard/ngx-forms-core";
+import {Directive, effect, forwardRef, inject, NgZone} from '@angular/core';
 import {MAT_FORM_FIELD, MatFormFieldControl} from "@angular/material/form-field";
-import {fromEvent, Subject} from "rxjs";
+import {fromEvent} from "rxjs";
 import {AutofillMonitor} from "@angular/cdk/text-field";
 import {takeUntilDestroyed} from "@angular/core/rxjs-interop";
+import {NgxFormFieldDirective} from "./ngx-form-field.directive";
+import {FocusOptions} from "@angular/cdk/a11y";
 
 @Directive({
   selector: 'input[ngxInput], textarea[ngxInput], [contentEditable][ngxInput]',
@@ -18,37 +16,13 @@ import {takeUntilDestroyed} from "@angular/core/rxjs-interop";
     '[class.mat-mdc-form-field-input-control]': '_inMdcFormField',
   }
 })
-export class NgxInputDirective<T> implements MatFormFieldControl<T | undefined>, OnDestroy {
-
-  static nextId = 0;
-  protected uid = `ngx-input-${NgxInputDirective.nextId++}`;
-
-  readonly idIn = input<string | undefined>(undefined, {alias: 'id'});
-  protected readonly _id = computed(() => this.idIn() ?? this.uid);
-  get id() {
-    return this._id()
-  }
-
-  readonly inputValue: ModelSignal<T | undefined> = model<T | undefined>(undefined);
-  readonly node: InputSignal<FormNode<T> | undefined> = input<FormNode<T>>();
-
-  private readonly _value = computed(() => {
-    const node = this.node();
-    if (!node) return this.inputValue();
-    return node.state();
-  });
-
-  readonly touched = output();
-
-  readonly ngControl = null;
-
-  private element = inject(ElementRef<HTMLElement>).nativeElement;
-  private autofillMonitor = inject(AutofillMonitor);
+export class NgxInputDirective<T> extends NgxFormFieldDirective<T> {
 
   protected _inMdcFormField = inject(MAT_FORM_FIELD, {optional: true}) != null;
-  private lastValue?: { val: T | undefined };
+  private autofillMonitor = inject(AutofillMonitor);
 
   constructor() {
+    super();
 
     this.autofillMonitor.monitor(this.element)
       .subscribe(x => this._autofilled.set(x.isAutofilled));
@@ -65,38 +39,11 @@ export class NgxInputDirective<T> implements MatFormFieldControl<T | undefined>,
       this.element.placeholder = this._placeholder();
     });
 
-    effect(() => {
-      const value = this._value();
-
-      if (this.lastValue) {
-        const lastValue = this.lastValue.val;
-        this.lastValue = undefined;
-        if (lastValue === value) return;
-      }
-
-      this.writeValue(value);
-    });
-
-    effect(() => {
-      this._id();
-      this._value();
-      this._placeholder();
-      this._focused();
-      this._empty();
-      this._shouldLabelFloat();
-      this._required();
-      this._disabled();
-      this._errorState();
-      this._autofilled();
-      this.update();
-    });
-
     const zone = inject(NgZone);
     zone.runOutsideAngular(() => {
       fromEvent<InputEvent>(this.element, 'input').pipe(takeUntilDestroyed()).subscribe(x => {
-        const value = this.readValue();
-        this.lastValue = {val: value};
-        zone.run(() => this.value = value);
+        const value = this.element.value ?? undefined;
+        zone.run(() => this.setValue(value));
       });
 
       fromEvent(this.element, 'focus').pipe(takeUntilDestroyed()).subscribe(() => zone.run(() => this.onFocus()));
@@ -104,122 +51,21 @@ export class NgxInputDirective<T> implements MatFormFieldControl<T | undefined>,
     });
   }
 
-  protected readValue(): T | undefined {
-    return this.element.value ?? undefined;
-  }
-
-  protected writeValue(value: T | undefined) {
-    this.element.value = value ?? null;
-  }
-
   ngOnDestroy() {
     this.autofillMonitor.stopMonitoring(this.element);
   }
 
-  get value() {
-    return this._value()
+  focus(options: FocusOptions | undefined): void {
+    this.element.focus(options);
   }
 
-  set value(value: T | undefined) {
-    this.node()?.setValue(value);
-    this.inputValue.set(value);
-  }
-
-  private readonly _stateChanges = new Subject<void>();
-  readonly stateChanges = this._stateChanges.asObservable();
-
-  protected update() {
-    this._stateChanges.next()
-  };
-
-  readonly placeholderIn: InputSignal<string | undefined> = input<string | undefined>(undefined, {alias: 'placeholder'})
-  protected readonly _placeholder = computed(() => this.placeholderIn() ?? '');
-  get placeholder() {
-    return this._placeholder()
-  };
-
-  protected readonly _focused = signal(false);
-  get focused() {
-    return this._focused()
-  };
-
-  protected readonly _empty = computed(() => !this._autofilled() && this.isEmpty(this._value()));
-  get empty() {
-    return this._empty()
-  };
-
-  protected isEmpty(value: T | undefined) {
+  protected isEmpty(value: T | undefined): boolean {
     if (value == null) return true;
     return value === '';
   }
 
-  protected readonly _shouldLabelFloat = computed(() => this._focused() || !this._empty())
-  get shouldLabelFloat() {
-    return this._shouldLabelFloat()
-  };
-
-  readonly requiredIn: InputSignalWithTransform<boolean, unknown> = input(false, {transform: booleanAttribute, alias: 'required'});
-  protected readonly _required = computed(() => this.requiredIn() || this.node()?.required || false);
-  get required() {
-    return this._required()
-  };
-
-  readonly disabledIn: InputSignalWithTransform<boolean, unknown> = input(false, {transform: booleanAttribute, alias: 'disabled'});
-  protected readonly _disabled = computed(() => this.disabledIn() || this.node()?.disabled() || false);
-  get disabled() {
-    return this._disabled()
-  };
-
-  readonly nodeErrorState = computed(() => {
-    const node = this.node();
-    if (!node) return false;
-    if (!node.hasError()) return false;
-    return node.touched() || node.changed();
-  });
-  readonly errorStateIn: InputSignalWithTransform<boolean, unknown> = input(false, {transform: booleanAttribute, alias: 'showError'});
-  protected readonly _errorState = computed(() => this.errorStateIn() || this.nodeErrorState());
-  get errorState() {
-    return this._errorState()
-  };
-
-  readonly controlType = 'ngx-input';
-
-  private readonly _autofilled = signal(false);
-  get autofilled() {
-    return this._autofilled()
-  };
-
-
-  readonly userAriaDescribedByIn: InputSignal<string | undefined> = input<string|undefined>(undefined, {alias: 'aria-describedby'});
-  protected readonly _userAriaDescribedBy = computed(() => this.userAriaDescribedByIn() ?? '');
-  get userAriaDescribedBy() {return this._userAriaDescribedBy()};
-
-  onContainerClick(event: MouseEvent): void {
-    if (this.focused) return;
-    this.focus();
+  protected writeValue(value: T | undefined): void {
+    this.element.value = value ?? null;
   }
 
-  setDescribedByIds(ids: string[]): void {
-    if (ids.length) {
-      this.element.setAttribute('aria-describedby', ids.join(' '));
-    } else {
-      this.element.removeAttribute('aria-describedby');
-    }
-  }
-
-  focus(options?: FocusOptions): void {
-    this.element.focus(options);
-  }
-
-  private onFocus() {
-    if (this.focused) return;
-    this._focused.set(true);
-  }
-
-  private onBlur() {
-    if (!this.focused) return;
-    this._focused.set(false);
-    this.node()?.markAsTouched();
-    this.touched.emit();
-  }
 }